+++ src/main/java/ru/mail/polis/lsm/artem_drozdov/MergeIterator.java
@@ -1,0 +1,50 @@
+package ru.mail.polis.lsm.artem_drozdov;
+
+import ru.mail.polis.lsm.Record;
+
+import java.nio.ByteBuffer;
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+
+public class MergeIterator implements Iterator<Record> {
+
+    private final LsmDAO.PeekingIterator left;
+    private final LsmDAO.PeekingIterator right;
+
+    public MergeIterator(LsmDAO.PeekingIterator left, LsmDAO.PeekingIterator right) {
+        this.left = left;
+        this.right = right;
+    }
+
+    @Override
+    public boolean hasNext() {
+        return left.hasNext() || right.hasNext();
+    }
+
+    @Override
+    public Record next() {
+        if (!hasNext()) {
+            throw new NoSuchElementException("No elements");
+        }
+
+        if (!left.hasNext()) {
+            return right.next();
+        }
+        if (!right.hasNext()) {
+            return left.next();
+        }
+
+        // checked earlier
+        ByteBuffer leftKey = Objects.requireNonNull(left.peek()).getKey();
+        ByteBuffer rightKey = Objects.requireNonNull(right.peek()).getKey();
+
+        int compareResult = leftKey.compareTo(rightKey);
+        if (compareResult == 0) {
+            left.next();
+            return right.next();
+        }
+
+        return compareResult < 0 ? left.next() : right.next();
+    }
+}
